requirements.txt
functions-framework==3.*
line-bot-sdk
elasticsearch
google-cloud-logging
icrawler
